requirements.txt
alembic==1.14.0
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
confluent-kafka==2.6.1
Flask==3.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kafka-python==2.0.2
Mako==1.3.8
MarkupSafe==2.1.4
mypy==1.13.0
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
mysqlclient==2.1.1
packaging==23.2
pkgutil_resolve_name==1.3.10
psycopg2==2.9.10
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
soupsieve==2.5
SQLAlchemy==2.0.27
tomli==2.2.1
types-beautifulsoup4==4.12.0.20241020
types-html5lib==1.1.11.20241018
types-requests==2.32.0.20241016
typing_extensions==4.9.0
urllib3==2.2.3
Werkzeug==3.0.1
zipp==3.17.0
